refactor code for reusability
