+++ apps/tenant-management-webapp/src/app/pages/admin/services/file/fileTypesTable.tsx
@@ -4,7 +4,8 @@ import { Role } from '@store/tenant/models';
 import { GoAContextMenu, GoAContextMenuIcon } from '@components/ContextMenu';
 import styled from 'styled-components';
 import { GoABadge, GoAButtonGroup, GoAModal, GoAButton } from '@abgov/react-components-new';
-import { FileTypeItem, SecurityClassifications } from '@store/file/models';
+import { FileTypeItem } from '@store/file/models';
+import { SecurityClassifications } from '@store/common/models';
 import { useHistory, useRouteMatch } from 'react-router-dom';
 import { DeleteModal } from '@components/DeleteModal';
 import { useDispatch } from 'react-redux';
