+++ libs/adsp-service-sdk/src/common/types.ts
@@ -1,0 +1,6 @@
+export enum SecurityClassifications {
+  Protected_A = 'Protected A',
+  Protected_B = 'Protected B',
+  Protected_C = 'Protected C',
+  Public = 'Public',
+}
